Make client a class without an interface
