+++ examples/restjson-example/src/main/java/software/amazon/smithy/java/runtime/example/PersonDirectoryAsyncClient.java
@@ -8,44 +8,57 @@ package software.amazon.smithy.java.runtime.example;
 import java.util.concurrent.CompletableFuture;
 import software.amazon.smithy.java.runtime.client.core.Client;
 import software.amazon.smithy.java.runtime.core.Context;
+import software.amazon.smithy.java.runtime.example.model.GetPersonImage;
 import software.amazon.smithy.java.runtime.example.model.GetPersonImageInput;
 import software.amazon.smithy.java.runtime.example.model.GetPersonImageOutput;
+import software.amazon.smithy.java.runtime.example.model.PutPerson;
+import software.amazon.smithy.java.runtime.example.model.PutPersonImage;
 import software.amazon.smithy.java.runtime.example.model.PutPersonImageInput;
 import software.amazon.smithy.java.runtime.example.model.PutPersonImageOutput;
 import software.amazon.smithy.java.runtime.example.model.PutPersonInput;
 import software.amazon.smithy.java.runtime.example.model.PutPersonOutput;
 
-public interface PersonDirectoryAsyncClient {
+public final class PersonDirectoryAsyncClient extends Client {
 
-    default CompletableFuture<GetPersonImageOutput> getPersonImage(GetPersonImageInput input) {
+    private PersonDirectoryAsyncClient(Builder builder) {
+        super(builder);
+    }
+
+    public CompletableFuture<GetPersonImageOutput> getPersonImage(GetPersonImageInput input) {
         return getPersonImage(input, Context.create());
     }
 
-    CompletableFuture<GetPersonImageOutput> getPersonImage(GetPersonImageInput input, Context context);
+    public CompletableFuture<GetPersonImageOutput> getPersonImage(GetPersonImageInput input, Context context) {
+        return call(input, null, null, new GetPersonImage(), context);
+    }
 
-    default CompletableFuture<PutPersonOutput> putPerson(PutPersonInput input) {
+    public CompletableFuture<PutPersonOutput> putPerson(PutPersonInput input) {
         return putPerson(input, Context.create());
     }
 
-    CompletableFuture<PutPersonOutput> putPerson(PutPersonInput input, Context context);
+    public CompletableFuture<PutPersonOutput> putPerson(PutPersonInput input, Context context) {
+        return call(input, null, null, new PutPerson(), context);
+    }
 
-    default CompletableFuture<PutPersonImageOutput> putPersonImage(PutPersonImageInput input) {
+    public CompletableFuture<PutPersonImageOutput> putPersonImage(PutPersonImageInput input) {
         return putPersonImage(input, Context.create());
     }
 
-    CompletableFuture<PutPersonImageOutput> putPersonImage(PutPersonImageInput input, Context context);
+    public CompletableFuture<PutPersonImageOutput> putPersonImage(PutPersonImageInput input, Context context) {
+        return call(input, input.image(), null, new PutPersonImage(), context);
+    }
 
-    static Builder builder() {
+    public static Builder builder() {
         return new Builder();
     }
 
-    final class Builder extends Client.Builder<PersonDirectoryAsyncClient, Builder> {
+    public static final class Builder extends Client.Builder<Builder> {
 
         private Builder() {}
 
         @Override
         public PersonDirectoryAsyncClient build() {
-            return new PersonDirectoryAsyncClientImpl(this);
+            return new PersonDirectoryAsyncClient(this);
         }
     }
 }
