+++ examples/restjson-example/src/main/java/software/amazon/smithy/java/runtime/example/PersonDirectoryClient.java
@@ -7,44 +7,57 @@ package software.amazon.smithy.java.runtime.example;
 
 import software.amazon.smithy.java.runtime.client.core.Client;
 import software.amazon.smithy.java.runtime.core.Context;
+import software.amazon.smithy.java.runtime.example.model.GetPersonImage;
 import software.amazon.smithy.java.runtime.example.model.GetPersonImageInput;
 import software.amazon.smithy.java.runtime.example.model.GetPersonImageOutput;
+import software.amazon.smithy.java.runtime.example.model.PutPerson;
+import software.amazon.smithy.java.runtime.example.model.PutPersonImage;
 import software.amazon.smithy.java.runtime.example.model.PutPersonImageInput;
 import software.amazon.smithy.java.runtime.example.model.PutPersonImageOutput;
 import software.amazon.smithy.java.runtime.example.model.PutPersonInput;
 import software.amazon.smithy.java.runtime.example.model.PutPersonOutput;
 
-public interface PersonDirectoryClient {
+public final class PersonDirectoryClient extends Client {
 
-    default GetPersonImageOutput getPersonImage(GetPersonImageInput input) {
+    private PersonDirectoryClient(Builder builder) {
+        super(builder);
+    }
+
+    public GetPersonImageOutput getPersonImage(GetPersonImageInput input) {
         return getPersonImage(input, Context.create());
     }
 
-    GetPersonImageOutput getPersonImage(GetPersonImageInput input, Context context);
+    public GetPersonImageOutput getPersonImage(GetPersonImageInput input, Context context) {
+        return call(input, null, null, new GetPersonImage(), context).join();
+    }
 
-    default PutPersonOutput putPerson(PutPersonInput input) {
+    public PutPersonOutput putPerson(PutPersonInput input) {
         return putPerson(input, Context.create());
     }
 
-    PutPersonOutput putPerson(PutPersonInput input, Context context);
+    public PutPersonOutput putPerson(PutPersonInput input, Context context) {
+        return call(input, null, null, new PutPerson(), context).join();
+    }
 
-    default PutPersonImageOutput putPersonImage(PutPersonImageInput input) {
+    public PutPersonImageOutput putPersonImage(PutPersonImageInput input) {
         return putPersonImage(input, Context.create());
     }
 
-    PutPersonImageOutput putPersonImage(PutPersonImageInput input, Context context);
-
-    static Builder builder() {
-        return new Builder();
+    public PutPersonImageOutput putPersonImage(PutPersonImageInput input, Context context) {
+        return call(input, input.image(), null, new PutPersonImage(), context).join();
     }
 
-    final class Builder extends Client.Builder<PersonDirectoryClient, Builder> {
+    public static PersonDirectoryClient.Builder builder() {
+        return new PersonDirectoryClient.Builder();
+    }
+
+    public static final class Builder extends Client.Builder<Builder> {
 
         private Builder() {}
 
         @Override
         public PersonDirectoryClient build() {
-            return new PersonDirectoryClientImpl(this);
+            return new PersonDirectoryClient(this);
         }
     }
 }
