--- examples/restjson-example/src/main/java/software/amazon/smithy/java/runtime/example/PersonDirectoryClientImpl.java
@@ -1,40 +1,0 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package software.amazon.smithy.java.runtime.example;
-
-import software.amazon.smithy.java.runtime.client.core.Client;
-import software.amazon.smithy.java.runtime.core.Context;
-import software.amazon.smithy.java.runtime.example.model.GetPersonImage;
-import software.amazon.smithy.java.runtime.example.model.GetPersonImageInput;
-import software.amazon.smithy.java.runtime.example.model.GetPersonImageOutput;
-import software.amazon.smithy.java.runtime.example.model.PutPerson;
-import software.amazon.smithy.java.runtime.example.model.PutPersonImage;
-import software.amazon.smithy.java.runtime.example.model.PutPersonImageInput;
-import software.amazon.smithy.java.runtime.example.model.PutPersonImageOutput;
-import software.amazon.smithy.java.runtime.example.model.PutPersonInput;
-import software.amazon.smithy.java.runtime.example.model.PutPersonOutput;
-
-final class PersonDirectoryClientImpl extends Client implements PersonDirectoryClient {
-
-    PersonDirectoryClientImpl(PersonDirectoryClient.Builder builder) {
-        super(builder);
-    }
-
-    @Override
-    public PutPersonOutput putPerson(PutPersonInput input, Context context) {
-        return call(input, null, null, new PutPerson(), context).join();
-    }
-
-    @Override
-    public PutPersonImageOutput putPersonImage(PutPersonImageInput input, Context context) {
-        return call(input, input.image(), null, new PutPersonImage(), context).join();
-    }
-
-    @Override
-    public GetPersonImageOutput getPersonImage(GetPersonImageInput input, Context context) {
-        return call(input, null, null, new GetPersonImage(), context).join();
-    }
-}
